+++ packages/wonder-blocks-dropdown/src/components/__tests__/dropdown-core.test.tsx
@@ -1,3 +1,4 @@
+// We allow raw buttons in this test since DropdownCore is a low-level component.
 import * as React from "react";
 import {fireEvent, render, screen, waitFor} from "@testing-library/react";
 import {userEvent} from "@testing-library/user-event";