+++ packages/wonder-blocks-popover/src/util/__tests__/util.test.tsx
@@ -1,3 +1,4 @@
+// Allowing raw buttons in an internal util file to test functionality
 import * as React from "react";
 import {render, screen} from "@testing-library/react";
 import {isFocusable} from "../util";