+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-link
 
+## 10.3.2
+
+### Patch Changes
+
+- Updated dependencies [d69335d]
+    - @khanacademy/wonder-blocks-clickable@8.2.2
+    - @khanacademy/wonder-blocks-core@12.4.4
+
 ## 10.3.1
 
 ### Patch Changes